requirements.txt
numpy==1.16.4
Pillow==6.1.0
pytest==5.0.1
torch==1.1.0
torchvision==0.3.0
